+++ queries/html/rainbow-tags.scm
@@ -3,10 +3,13 @@
 
 (element
   ((start_tag) @opening
-   ; (element (self_closing_tag) @intermediate)?
+   ; (element (self_closing_tag) @intermediate)*
    (end_tag) @closing)) @container
 
 (style_element
   ((start_tag) @opening
-   ; (element (self_closing_tag) @intermediate)?
+   (end_tag) @closing)) @container
+
+(script_element
+  ((start_tag) @opening
    (end_tag) @closing)) @container